Push external/function code out of decorators.py and into the scheduler (#1173)
Inspired by #1171 - this pushes all the code from `function.__call__` and `external.__call__` into `scheduler.run_in_executor` and `scheduler.queue_function`.

Neither `scheduler.run_in_executor` nor `scheduler.queue_function` were remotely safe to use directly anyway, and were heavily entangled with the design of the decorators.

After this change, the decorators become nothing but syntactic sugar.
