+++ cocotb/scheduler.py
@@ -62,7 +62,7 @@ import cocotb.decorators
 from cocotb.triggers import (Trigger, GPITrigger, Timer, ReadOnly,
                              NextTimeStep, ReadWrite, Event, Join, NullTrigger)
 from cocotb.log import SimLog
-from cocotb.result import TestComplete
+from cocotb.result import TestComplete, ReturnValue
 from cocotb import _py_compat
 
 # On python 3.7 onwards, `dict` is guaranteed to preserve insertion order.
@@ -561,7 +561,7 @@ class Scheduler(object):
         """Queue a coroutine for execution"""
         self._pending_coros.append(coroutine)
 
-    def queue_function(self, coroutine):
+    def queue_function(self, coro):
         """Queue a coroutine for execution and move the containing thread
         so that it does not block execution of the main thread any longer.
         """
@@ -578,9 +578,22 @@ class Scheduler(object):
         # each entry always has a unique thread.
         t, = matching_threads
 
+        @cocotb.coroutine
+        def wrapper():
+            try:
+                _outcome = outcomes.Value((yield coro))
+            except BaseException as e:
+                _outcome = outcomes.Error(e)
+            event.outcome = _outcome
+            event.set()
+
+        event = threading.Event()
         t.thread_suspend()
-        self._pending_coros.append(coroutine)
-        return t
+        self._pending_coros.append(wrapper())
+        # This blocks the calling external thread until the coroutine finishes
+        event.wait()
+        t.thread_resume()
+        return event.outcome.get()
 
     def run_in_executor(self, func, *args, **kwargs):
         """Run the coroutine in a separate execution thread
@@ -598,15 +611,22 @@ class Scheduler(object):
                 self.log.debug("Execution of external routine done %s" % threading.current_thread())
             _waiter.thread_done()
 
-        waiter = external_waiter()
-        thread = threading.Thread(group=None, target=execute_external,
-                                  name=func.__name__ + "_thread",
-                                  args=([func, waiter]), kwargs={})
+        @cocotb.coroutine
+        def wrapper():
+            waiter = external_waiter()
+            thread = threading.Thread(group=None, target=execute_external,
+                                      name=func.__name__ + "_thread",
+                                      args=([func, waiter]), kwargs={})
 
-        waiter.thread = thread
-        self._pending_threads.append(waiter)
+            waiter.thread = thread
+            self._pending_threads.append(waiter)
 
-        return waiter
+            yield waiter.event.wait()
+
+            ret = waiter.result  # raises if there was an exception
+            raise ReturnValue(ret)
+
+        return wrapper()
 
     def add(self, coroutine):
         """Add a new coroutine.
